+++ src/Controller/RequestController.php
@@ -2,9 +2,12 @@
 
 namespace App\Controller;
 
+use App\Entity\Examination;
 use App\Entity\Request as EntityRequest;
 use App\Form\RequestType;
 use App\Entity\User;
+use App\Form\ExaminationType;
+use App\Repository\ExaminationRepository;
 use App\Repository\RequestRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,14 +17,32 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 class RequestController extends AbstractController
 {
+    private RequestRepository $rr;
+    private ExaminationRepository $er;
+
+    public function __construct(
+        RequestRepository $rr,
+        ExaminationRepository $er
+    ) {
+        $this->rr = $rr;
+        $this->er = $er;
+    }
+
     #[Route('/admin/requests', name: 'requests')]
-    public function requests(Request $request, RequestRepository $rr): Response
+    public function requests(Request $request): Response
     {
+        $vet = null;
+        if (!$this->isGranted('ROLE_CARER')) {
+            /** @var User */
+            $user = $this->getUser();
+            $vet = $user->getVeterinary();
+        }
+
         $filter = $request->query->get('filter', 'waiting');
         $requests = match ($filter) {
-            'scheduled' => $rr->findAllScheduled(),
-            'done' => $rr->findAllDone(),
-            default => $rr->findAllWaiting(),
+            'scheduled' => $this->rr->findAllScheduled($vet),
+            'done' => $this->rr->findAllDone($vet),
+            default => $this->rr->findAllWaiting($vet),
         };
 
         return $this->render('request/requests.html.twig', [
@@ -30,11 +51,18 @@ class RequestController extends AbstractController
         ]);
     }
 
-    #[Route('/admin/request/editor', name: 'request_editor')]
+    #[Route('/admin/request/editor/{id?}', name: 'request_editor')]
     #[IsGranted('ROLE_CARER')]
-    public function editor(Request $request, RequestRepository $rr): Response
+    public function editor(Request $request, int|null $id): Response
     {
         $req = new EntityRequest();
+        if ($id) {
+            $req = $this->rr->findOneBy(['id' => $id]);
+        }
+
+        if (!$req) {
+            throw $this->createNotFoundException();
+        }
 
         $form = $this->createForm(RequestType::class, $req);
         $form->handleRequest($request);
@@ -45,7 +73,7 @@ class RequestController extends AbstractController
             $req
                 ->setCaregiver($user->getCaregiver())
                 ->setDateCreated(new \DateTime());
-            $rr->save($req);
+            $this->rr->save($req);
             return $this->redirectToRoute('requests');
         }
 
@@ -55,15 +83,53 @@ class RequestController extends AbstractController
     }
 
     #[Route('/admin/request/{id}', name: 'request')]
-    public function request(RequestRepository $rr, int $id): Response
+    public function request(int $id): Response
     {
-        $request = $rr->findOneBy(['id' => $id]);
+        /** @var User */
+        $user = $this->getUser();
+        $request = $this->rr->findOneBy(['id' => $id]);
         if (!$request) {
             return $this->redirectToRoute('requests');
         }
 
         return $this->render('request/index.html.twig', [
             'request' => $request,
+            'vet' => $user->getVeterinary() == $request->getVeterinary(),
+        ]);
+    }
+
+    #[Route('/admin/examination/{id}', name: 'exam_editor')]
+    #[IsGranted('ROLE_VET')]
+    public function exam_editor(Request $request, int $id): Response
+    {
+        $req = new EntityRequest();
+        if ($id) {
+            $req = $this->rr->findOneBy(['id' => $id]);
+        }
+
+        if (!$req) {
+            throw $this->createNotFoundException();
+        }
+
+        $exam = $req->getExamination() ?? new Examination();
+        $form = $this->createForm(ExaminationType::class, $exam);
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            /** @var User */
+            $user = $this->getUser();
+
+            $exam->setRequest($req);
+            $exam->setVeterinary($user->getVeterinary());
+            $exam->setAnimal($req->getAnimal());
+            $exam->setResult($exam->getResult() ?? '');
+            $this->er->save($exam);
+
+            return $this->redirectToRoute('requests');
+        }
+
+        return $this->render('request/exam_editor.html.twig', [
+            'request' => $req,
+            'form' => $form,
         ]);
     }
 }